Rename: WidgetModal1.vue -> WidgetModal.vue
