+++ frontend/src/components/WidgetModal.vue
